[rename] JwtUtilTest 위치 변경
- 본 코드와 테스트코드가 같은 위치에 위치하도록 디렉토리 변경

Related to: #30
